+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#20](https://github.com/LaxarJS/laxar-tooling/issues/20): stop including multiple theme assets in artifact listing
+
 
 ## v0.5.0-alpha.5